requirements.txt
python-telegram-bot==21.0
python-dotenv==1.0.0
requests>=2.25.0
loguru==0.7.2
telethon>=1.28.0
yookassa>=3.0.0
# psycopg2-binary==2.9.9  # Временно отключено для тестирования 